+++ src/builder_context.rs
@@ -1,8 +1,9 @@
 use crate::{
-    ActiveEnumConfig, ActiveEnumFilterInputConfig, ConnectionObjectConfig, CursorInputConfig,
-    EdgeObjectConfig, EntityCreateBatchMutationConfig, EntityCreateOneMutationConfig,
-    EntityDeleteMutationConfig, EntityInputConfig, EntityObjectConfig, EntityQueryFieldConfig,
-    EntityUpdateMutationConfig, FilterInputConfig, OffsetInputConfig, OrderByEnumConfig,
+    ActiveEnumConfig, ActiveEnumFilterInputConfig, CascadeInputConfig, ConnectionObjectConfig,
+    CursorInputConfig, EdgeObjectConfig, EntityCreateBatchMutationConfig,
+    EntityCreateOneMutationConfig, EntityDeleteMutationConfig, EntityGetFieldConfig,
+    EntityInputConfig, EntityObjectConfig, EntityQueryFieldConfig, EntityUpdateMutationConfig,
+    FilterInputConfig, NewOrderInputConfig, OffsetInputConfig, OrderByEnumConfig, OrderEnumConfig,
     OrderInputConfig, PageInfoObjectConfig, PageInputConfig, PaginationInfoObjectConfig,
     PaginationInputConfig,
 };
@@ -23,6 +24,7 @@ pub use filter_types_map::*;
 #[derive(Default)]
 pub struct BuilderContext {
     pub order_by_enum: OrderByEnumConfig,
+    pub order_enum: OrderEnumConfig,
     pub active_enum: ActiveEnumConfig,
 
     pub cursor_input: CursorInputConfig,
@@ -31,8 +33,10 @@ pub struct BuilderContext {
     pub pagination_input: PaginationInputConfig,
 
     pub order_input: OrderInputConfig,
+    pub new_order_input: NewOrderInputConfig,
 
     pub filter_input: FilterInputConfig,
+    pub cascade_input: CascadeInputConfig,
     pub active_enum_filter_input: ActiveEnumFilterInputConfig,
 
     pub page_info_object: PageInfoObjectConfig,
@@ -41,6 +45,7 @@ pub struct BuilderContext {
     pub entity_object: EntityObjectConfig,
     pub connection_object: ConnectionObjectConfig,
     pub entity_query_field: EntityQueryFieldConfig,
+    pub entity_get_field: EntityGetFieldConfig,
 
     pub entity_create_one_mutation: EntityCreateOneMutationConfig,
     pub entity_create_batch_mutation: EntityCreateBatchMutationConfig,