+++ src/enumerations/active_enum.rs
@@ -1,5 +1,5 @@
 use async_graphql::dynamic::Enum;
-use heck::ToUpperCamelCase;
+use heck::{ToSnakeCase, ToUpperCamelCase};
 use sea_orm::{ActiveEnum, DynIden, Value};
 
 use crate::BuilderContext;
@@ -19,7 +19,16 @@ impl std::default::Default for ActiveEnumConfig {
                 format!("{}Enum", name.to_upper_camel_case())
             }),
             variant_name: Box::new(|_enum_name: &str, variant: &str| -> String {
-                variant.to_upper_camel_case().to_ascii_uppercase()
+                if cfg!(feature = "field-snake-case") {
+                    variant.to_snake_case()
+                } else if cfg!(feature = "offset-pagination") {
+                    variant
+                        .chars()
+                        .filter(|c| c.is_ascii_alphanumeric())
+                        .collect()
+                } else {
+                    variant.to_upper_camel_case().to_ascii_uppercase()
+                }
             }),
         }
     }