+++ src/enumerations/mod.rs
@@ -3,3 +3,6 @@ pub use order_by_enum::*;
 
 pub mod active_enum;
 pub use active_enum::*;
+
+pub mod order_enum;
+pub use order_enum::*;