+++ src/inputs/active_enum_filter_input.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeSet;
 
 use async_graphql::dynamic::ObjectAccessor;
-use heck::ToUpperCamelCase;
+use heck::{ToSnakeCase, ToUpperCamelCase};
 use sea_orm::{ActiveEnum, ColumnTrait, ColumnType, Condition, DynIden, EntityTrait};
 
 use crate::{ActiveEnumBuilder, BuilderContext, FilterInfo, FilterOperation, SeaResult};
@@ -88,10 +88,18 @@ where
 
     let extract_variant = move |input: &str| -> String {
         let variant = variants.iter().find(|variant| {
-            let variant = variant
-                .to_string()
-                .to_upper_camel_case()
-                .to_ascii_uppercase();
+            let variant = variant.to_string();
+            let variant = if cfg!(feature = "field-snake-case") {
+                variant.to_snake_case()
+            } else if cfg!(feature = "offset-pagination") {
+                variant
+                    .chars()
+                    .filter(|c| c.is_alphanumeric())
+                    .collect::<String>()
+            } else {
+                variant.to_upper_camel_case().to_ascii_uppercase()
+            };
+
             variant.eq(input)
         });
         variant.unwrap().to_string()
@@ -139,7 +147,7 @@ where
         condition
     };
 
-    let condition = match filter.get("is_in") {
+    let condition = match filter.get("in") {
         Some(data) => {
             let data: Vec<_> = data
                 .list()