+++ src/query/entity_object_relation.rs
@@ -1,17 +1,18 @@
 use async_graphql::{
     dataloader::DataLoader,
-    dynamic::{Field, FieldFuture, FieldValue, InputValue, TypeRef},
+    dynamic::{Field, FieldFuture, FieldValue, InputValue, TypeRef, ValueAccessor},
     Error,
 };
-use heck::ToSnakeCase;
+use heck::{ToLowerCamelCase, ToSnakeCase};
 use sea_orm::{EntityTrait, Iden, ModelTrait, RelationDef};
 
 #[cfg(not(feature = "offset-pagination"))]
 use crate::ConnectionObjectBuilder;
 use crate::{
     apply_memory_pagination, get_filter_conditions, BuilderContext, EntityObjectBuilder,
-    FilterInputBuilder, GuardAction, HashableGroupKey, KeyComplex, OneToManyLoader, OneToOneLoader,
-    OrderInputBuilder, PaginationInputBuilder,
+    FilterInputBuilder, GuardAction, HashableGroupKey, KeyComplex, NewOrderInputBuilder,
+    OffsetInput, OneToManyLoader, OneToOneLoader, OrderInputBuilder, PageInput, PaginationInput,
+    PaginationInputBuilder,
 };
 
 /// This builder produces a GraphQL field for an SeaORM entity relationship
@@ -31,12 +32,18 @@ impl EntityObjectRelationBuilder {
         <R as sea_orm::EntityTrait>::Model: Sync,
         <<R as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
     {
+        let name = if cfg!(feature = "field-snake-case") {
+            name.to_snake_case()
+        } else {
+            name.to_lower_camel_case()
+        };
         let context: &'static BuilderContext = self.context;
         let entity_object_builder = EntityObjectBuilder { context };
         #[cfg(not(feature = "offset-pagination"))]
         let connection_object_builder = ConnectionObjectBuilder { context };
         let filter_input_builder = FilterInputBuilder { context };
         let order_input_builder = OrderInputBuilder { context };
+        let new_order_input_builder = NewOrderInputBuilder { context };
 
         let object_name: String = entity_object_builder.type_name::<R>();
         #[cfg(feature = "offset-pagination")]
@@ -100,7 +107,11 @@ impl EntityObjectRelationBuilder {
                     let filters = ctx.args.get(&context.entity_query_field.filters);
                     let filters = get_filter_conditions::<R>(context, filters);
                     let order_by = ctx.args.get(&context.entity_query_field.order_by);
-                    let order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let mut order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let order = ctx.args.get(&context.entity_query_field.order);
+                    let order = NewOrderInputBuilder { context }.parse_object::<R>(order);
+                    order_by.extend(order);
+
                     let key = KeyComplex::<R> {
                         key: vec![parent.get(from_col)],
                         meta: HashableGroupKey::<R> {
@@ -151,7 +162,10 @@ impl EntityObjectRelationBuilder {
                     let filters = ctx.args.get(&context.entity_query_field.filters);
                     let filters = get_filter_conditions::<R>(context, filters);
                     let order_by = ctx.args.get(&context.entity_query_field.order_by);
-                    let order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let mut order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let order = ctx.args.get(&context.entity_query_field.order);
+                    let order = NewOrderInputBuilder { context }.parse_object::<R>(order);
+                    order_by.extend(order);
                     let key = KeyComplex::<R> {
                         key: vec![parent.get(from_col)],
                         meta: HashableGroupKey::<R> {
@@ -165,6 +179,43 @@ impl EntityObjectRelationBuilder {
                     let values = loader.load_one(key).await?;
                     let pagination = ctx.args.get(&context.entity_query_field.pagination);
                     let pagination = PaginationInputBuilder { context }.parse_object(pagination);
+                    let first = ctx.args.get("first");
+                    let pagination = match first {
+                        Some(first_value) => match first_value.u64() {
+                            Ok(first_num) => {
+                                if let Some(offset) = pagination.offset {
+                                    PaginationInput {
+                                        offset: Some(OffsetInput {
+                                            offset: offset.offset,
+                                            limit: first_num,
+                                        }),
+                                        page: None,
+                                        cursor: None,
+                                    }
+                                } else if let Some(page) = pagination.page {
+                                    PaginationInput {
+                                        offset: None,
+                                        page: Some(PageInput {
+                                            page: page.page,
+                                            limit: first_num,
+                                        }),
+                                        cursor: None,
+                                    }
+                                } else {
+                                    PaginationInput {
+                                        offset: Some(OffsetInput {
+                                            offset: 0,
+                                            limit: first_num,
+                                        }),
+                                        page: None,
+                                        cursor: None,
+                                    }
+                                }
+                            }
+                            _error => pagination,
+                        },
+                        None => pagination,
+                    };
 
                     let object = apply_memory_pagination::<R>(values, pagination);
 
@@ -173,21 +224,40 @@ impl EntityObjectRelationBuilder {
             }),
         };
 
-        match relation_definition.is_owner {
-            false => field,
-            true => field
-                .argument(InputValue::new(
-                    &context.entity_query_field.filters,
-                    TypeRef::named(filter_input_builder.type_name(&object_name)),
-                ))
-                .argument(InputValue::new(
-                    &context.entity_query_field.order_by,
-                    TypeRef::named(order_input_builder.type_name(&object_name)),
-                ))
-                .argument(InputValue::new(
-                    &context.entity_query_field.pagination,
-                    TypeRef::named(&context.pagination_input.type_name),
-                )),
-        }
+        field
+            .argument(InputValue::new(
+                &context.entity_query_field.filters,
+                TypeRef::named(filter_input_builder.type_name(&object_name)),
+            ))
+            .argument(InputValue::new(
+                &context.entity_query_field.order_by,
+                TypeRef::named(order_input_builder.type_name(&object_name)),
+            ))
+            .argument(InputValue::new(
+                &self.context.entity_query_field.order,
+                TypeRef::named(new_order_input_builder.type_name(&object_name)),
+            ))
+            .argument(InputValue::new(
+                &context.entity_query_field.pagination,
+                TypeRef::named(&context.pagination_input.type_name),
+            ))
+            .argument(InputValue::new("first", TypeRef::named(TypeRef::INT)))
+    }
+
+    pub fn joiin<T, R>(
+        &self,
+        relation_definition: RelationDef,
+        filter: Option<ValueAccessor>,
+    ) -> RelationDef
+    where
+        T: EntityTrait,
+        <T as EntityTrait>::Model: Sync,
+        <<T as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
+        R: EntityTrait,
+        <R as sea_orm::EntityTrait>::Model: Sync,
+        <<R as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
+    {
+        let filters = get_filter_conditions::<R>(self.context, filter);
+        relation_definition.on_condition(move |_left, _right| filters.to_owned())
     }
 }