+++ src/query/entity_object_via_relation.rs
@@ -1,11 +1,12 @@
 use async_graphql::{
     dataloader::DataLoader,
-    dynamic::{Field, FieldFuture, FieldValue, InputValue, TypeRef},
+    dynamic::{Field, FieldFuture, FieldValue, InputValue, TypeRef, ValueAccessor},
     Error,
 };
-use heck::ToSnakeCase;
+use heck::{ToLowerCamelCase, ToSnakeCase};
 use sea_orm::{
-    ColumnTrait, Condition, DatabaseConnection, EntityTrait, Iden, ModelTrait, QueryFilter, Related,
+    ColumnTrait, Condition, DatabaseConnection, EntityTrait, Iden, ModelTrait, QueryFilter,
+    Related, RelationDef,
 };
 
 #[cfg(not(feature = "offset-pagination"))]
@@ -13,7 +14,8 @@ use crate::ConnectionObjectBuilder;
 use crate::{
     apply_memory_pagination, apply_order, apply_pagination, get_filter_conditions, BuilderContext,
     EntityObjectBuilder, FilterInputBuilder, GuardAction, HashableGroupKey, KeyComplex,
-    OneToManyLoader, OneToOneLoader, OrderInputBuilder, PaginationInputBuilder,
+    NewOrderInputBuilder, OffsetInput, OneToManyLoader, OneToOneLoader, OrderInputBuilder,
+    PageInput, PaginationInput, PaginationInputBuilder,
 };
 
 /// This builder produces a GraphQL field for an SeaORM entity related trait
@@ -34,6 +36,11 @@ impl EntityObjectViaRelationBuilder {
         <<T as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
         <<R as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
     {
+        let name = if cfg!(feature = "field-snake-case") {
+            name.to_snake_case()
+        } else {
+            name.to_lower_camel_case()
+        };
         let context: &'static BuilderContext = self.context;
         let to_relation_definition = <T as Related<R>>::to();
         let (via_relation_definition, is_via_relation) = match <T as Related<R>>::via() {
@@ -46,6 +53,7 @@ impl EntityObjectViaRelationBuilder {
         let connection_object_builder = ConnectionObjectBuilder { context };
         let filter_input_builder = FilterInputBuilder { context };
         let order_input_builder = OrderInputBuilder { context };
+        let new_order_input_builder = NewOrderInputBuilder { context };
         let object_name: String = entity_object_builder.type_name::<R>();
         #[cfg(feature = "offset-pagination")]
         let type_ref = TypeRef::named_list(&object_name);
@@ -113,7 +121,10 @@ impl EntityObjectViaRelationBuilder {
                     let filters = ctx.args.get(&context.entity_query_field.filters);
                     let filters = get_filter_conditions::<R>(context, filters);
                     let order_by = ctx.args.get(&context.entity_query_field.order_by);
-                    let order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let mut order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let order = ctx.args.get(&context.entity_query_field.order);
+                    let order = NewOrderInputBuilder { context }.parse_object::<R>(order);
+                    order_by.extend(order);
                     let key = KeyComplex::<R> {
                         key: vec![parent.get(from_col)],
                         meta: HashableGroupKey::<R> {
@@ -170,11 +181,51 @@ impl EntityObjectViaRelationBuilder {
                     let filters = get_filter_conditions::<R>(context, filters);
 
                     let order_by = ctx.args.get(&context.entity_query_field.order_by);
-                    let order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+                    let mut order_by = OrderInputBuilder { context }.parse_object::<R>(order_by);
+
+                    let order = ctx.args.get(&context.entity_query_field.order);
+                    let order = NewOrderInputBuilder { context }.parse_object::<R>(order);
+                    order_by.extend(order);
 
                     let pagination = ctx.args.get(&context.entity_query_field.pagination);
                     let pagination = PaginationInputBuilder { context }.parse_object(pagination);
-
+                    let first = ctx.args.get("first");
+                    let pagination = match first {
+                        Some(first_value) => match first_value.u64() {
+                            Ok(first_num) => {
+                                if let Some(offset) = pagination.offset {
+                                    PaginationInput {
+                                        offset: Some(OffsetInput {
+                                            offset: offset.offset,
+                                            limit: first_num,
+                                        }),
+                                        page: None,
+                                        cursor: None,
+                                    }
+                                } else if let Some(page) = pagination.page {
+                                    PaginationInput {
+                                        offset: None,
+                                        page: Some(PageInput {
+                                            page: page.page,
+                                            limit: first_num,
+                                        }),
+                                        cursor: None,
+                                    }
+                                } else {
+                                    PaginationInput {
+                                        offset: Some(OffsetInput {
+                                            offset: 0,
+                                            limit: first_num,
+                                        }),
+                                        page: None,
+                                        cursor: None,
+                                    }
+                                }
+                            }
+                            _error => pagination,
+                        },
+                        None => pagination,
+                    };
                     let db = ctx.data::<DatabaseConnection>()?;
 
                     let object = if is_via_relation {
@@ -206,21 +257,40 @@ impl EntityObjectViaRelationBuilder {
             }),
         };
 
-        match via_relation_definition.is_owner {
-            false => field,
-            true => field
-                .argument(InputValue::new(
-                    &context.entity_query_field.filters,
-                    TypeRef::named(filter_input_builder.type_name(&object_name)),
-                ))
-                .argument(InputValue::new(
-                    &context.entity_query_field.order_by,
-                    TypeRef::named(order_input_builder.type_name(&object_name)),
-                ))
-                .argument(InputValue::new(
-                    &context.entity_query_field.pagination,
-                    TypeRef::named(&context.pagination_input.type_name),
-                )),
-        }
+        field
+            .argument(InputValue::new(
+                &context.entity_query_field.filters,
+                TypeRef::named(filter_input_builder.type_name(&object_name)),
+            ))
+            .argument(InputValue::new(
+                &context.entity_query_field.order_by,
+                TypeRef::named(order_input_builder.type_name(&object_name)),
+            ))
+            .argument(InputValue::new(
+                &self.context.entity_query_field.order,
+                TypeRef::named(new_order_input_builder.type_name(&object_name)),
+            ))
+            .argument(InputValue::new(
+                &context.entity_query_field.pagination,
+                TypeRef::named(&context.pagination_input.type_name),
+            ))
+            .argument(InputValue::new("first", TypeRef::named(TypeRef::INT)))
+    }
+
+    pub fn joiin<T, R>(
+        &self,
+        relation_definition: RelationDef,
+        filter: Option<ValueAccessor>,
+    ) -> RelationDef
+    where
+        T: EntityTrait,
+        <T as EntityTrait>::Model: Sync,
+        <<T as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
+        R: EntityTrait,
+        <R as sea_orm::EntityTrait>::Model: Sync,
+        <<R as sea_orm::EntityTrait>::Column as std::str::FromStr>::Err: core::fmt::Debug,
+    {
+        let filters = get_filter_conditions::<R>(self.context, filter);
+        relation_definition.on_condition(move |_left, _right| filters.to_owned())
     }
 }